Added background tracing, hw counters, et al

+++ public/client/TracyScoped.hpp
@@ -220,7 +220,7 @@ private:
     const bool m_active;
 
 #ifdef TRACY_ON_DEMAND
-    uint64_t m_connectionId = 0;
+	uint32_t m_connectionId = 0;
 #endif
 };
